// SGRA-WEB/src/enums/bugEnum.ts
export enum BugType {
    FUNCIONAL = "FUNCIONAL",
    UI = "UI",
    RENDIMIENTO = "RENDIMIENTO",
    SEGURIDAD = "SEGURIDAD"
}

export enum BugSeverity {
    BAJO = "BAJO",
    MEDIO = "MEDIO",
    ALTO = "ALTO",
    SEVERO = "SEVERO",
    CRITICO = "CRITICO"
}

export enum BugStatus {
    PENDIENTE = "PENDIENTE",
    ASIGNADO = "ASIGNADO",
    EN_DESARROLLO = "EN_DESARROLLO",
    EN_REVISION = "EN_REVISION",
    ATRASADO = "ATRASADO",
    RESUELTO = "RESUELTO",
    CANCELADO = "CANCELADO"
}